add UserFactory;add PatchUserRequest and PATCH request to /users/{userId};fix 'last page' link in PaginationInfo;add 'enabled' field to User;now imposible to signin as unabled User

+++ src/main/java/it/sevenbits/taskmanager/core/repository/tasks/DatabaseTaskRepository.java
@@ -28,6 +28,7 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
     private final String taskStatus = "status";
     private final String taskCreatedAt = "createdAt";
     private final String taskUpdatedAt = "updatedAt";
+    private final String taskOwner = "owner";
 
     /**
      * Constructor for class
@@ -57,9 +58,10 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
                         String resultStatus = resultSet.getString(taskStatus);
                         String resultCreatedAt = resultSet.getString(taskCreatedAt);
                         String resultChangedAt = resultSet.getString(taskUpdatedAt);
+
                         return taskFactory.getNewTask(
                                 resultId, resultName, TaskStatus.resolveString(resultStatus),
-                                resultCreatedAt, resultChangedAt);
+                                resultCreatedAt, resultChangedAt, null);
                     },
                     status);
         } catch (Exception e) {
@@ -73,17 +75,22 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
      * @param order order of sorting resulting list
      * @param page number of page to return
      * @param size how much elements will be on page
+     * @param owner owner of tasks to get
      * @return List of Task objects of current page
      */
 
     public GetTasksResponse getTaskList(final String status, final String order, final Integer page,
-                                  final Integer size) {
-        int totalSize = getResultSize(status);
+                                  final Integer size, final String owner) {
+        if (owner == null) {
+            logger.warn("owner of taskList shouldn't be null");
+            return null;
+        }
+        int totalSize = getResultSize(status, owner);
         logger.debug("Total size of taskList without pagination- {}", totalSize);
         try {
             return new GetTasksResponse(jdbcOperations.query(
-                    "SELECT id, name, status, createdAt, updatedAt " +
-                            "FROM task WHERE status = ? ORDER by createdAt DESC LIMIT ? OFFSET ?",
+                    "SELECT id, name, status, createdAt, updatedAt, owner " +
+                            "FROM task WHERE status = ? AND owner = ? ORDER by createdAt DESC LIMIT ? OFFSET ?",
                     (resultSet, i) -> {
                         String resultId = resultSet.getString(taskId);
                         String resultName = resultSet.getString(taskText);
@@ -92,9 +99,9 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
                         String resultUpdatedAt = resultSet.getString(taskUpdatedAt);
                         return taskFactory.getNewTask(
                                 resultId, resultName, TaskStatus.resolveString(resultStatus),
-                                        resultCreatedAt, resultUpdatedAt);
+                                        resultCreatedAt, resultUpdatedAt, null);
                     },
-                    status, size, size * (page - 1)),
+                    status, owner, size, size * (page - 1)),
                     status, order, page, size, totalSize);
         } catch (Exception e) {
             logger.error(e.getMessage());
@@ -105,22 +112,27 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
     /**
      *Create new Task
      * @param text text of new Task
+     * @param owner owner of task
      * @return new Task Object or empty task if some error
      */
 
-    public Task createTask(final String text) {
+    public Task createTask(final String text, final String owner) {
+        if (owner == null) {
+            logger.warn("owner of task shouldn't be null");
+            return null;
+        }
         if (text == null || "".equals(text.trim())) {
             logger.warn("text of task to create shouldn't be empty");
             return null;
         }
         String id = getNewId().toString();
         TaskStatus status = TaskStatus.inbox;
-        Task result = taskFactory.getNewTask(id, text, status);
+        Task result = taskFactory.getNewTask(id, text, status, owner);
         try {
             int rows = jdbcOperations.update(
-                    "INSERT INTO task (id, name, status, createdAt, updatedAt) VALUES (?, ?, ?, ?, ?)",
+                    "INSERT INTO task (id, name, status, createdAt, updatedAt, owner) VALUES (?, ?, ?, ?, ?, ?)",
                     id, text, status.toString(), Timestamp.valueOf(result.getCreatedAt()),
-                    Timestamp.valueOf(result.getUpdatedAt())
+                    Timestamp.valueOf(result.getUpdatedAt()), result.getOwner()
             );
             if (rows > 0) {
                 return result;
@@ -150,18 +162,19 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
     public Task getTask(final String id) {
         try {
             return jdbcOperations.queryForObject(
-                    "SELECT id, name, status, createdAt, updatedAt FROM task WHERE id = ?",
+                    "SELECT id, name, status, createdAt, updatedAt,owner FROM task WHERE id = ?",
                     (resultSet, i) -> {
                         String rowId = resultSet.getString(taskId);
                         String rowName = resultSet.getString(taskText);
                         TaskStatus rowStatus = TaskStatus.resolveString(resultSet.getString(taskStatus));
                         String rowCreatedAt = resultSet.getString(taskCreatedAt);
                         String rowChangedAt = resultSet.getString(taskUpdatedAt);
+                        String rowOwner = resultSet.getString(taskOwner);
                         if (rowStatus == null) {
                             return null;
                         }
                         return taskFactory.getNewTask(rowId, rowName, rowStatus,
-                                rowCreatedAt, rowChangedAt);
+                                rowCreatedAt, rowChangedAt, rowOwner);
                     },
                     id);
         } catch (Exception e) {
@@ -179,15 +192,16 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
     public Task deleteTask(final String id) {
         try {
             Task deletedTask = jdbcOperations.queryForObject(
-                    "DELETE FROM task WHERE id = ? RETURNING id, name, status, createdAt, updatedAt",
+                    "DELETE FROM task WHERE id = ? RETURNING id, name, status, createdAt, updatedAt, owner",
                     (resultSet, i) -> {
                         TaskStatus rowStatus = TaskStatus.resolveString(resultSet.getString(taskStatus));
                         String rowId = resultSet.getString(taskId);
                         String rowName = resultSet.getString(taskText);
                         String rowCreatedAt = resultSet.getString(taskCreatedAt);
                         String rowChangedAt = resultSet.getString(taskUpdatedAt);
+                        String rowOwner = resultSet.getString(taskOwner);
                         return taskFactory.getNewTask(
-                                rowId, rowName, rowStatus, rowCreatedAt, rowChangedAt);
+                                rowId, rowName, rowStatus, rowCreatedAt, rowChangedAt, rowOwner);
                     },
                     id);
             if (deletedTask != null) {
@@ -209,11 +223,11 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
         if (changedTask != null) {
             try {
                 int rowsInsert = jdbcOperations.update(
-                        "INSERT INTO task (id, name, status, createdAt, updatedAt) VALUES (?, ?, ?, ?, ?) " +
+                        "INSERT INTO task (id, name, status, createdAt, updatedAt, owner) VALUES (?, ?, ?, ?, ?, ?) " +
                                 "ON CONFLICT(id) DO UPDATE SET name = ?, status = ?, updatedAt = ?",
                         changedTask.getId(), changedTask.getText(), changedTask.getStatus().toString(),
                         Timestamp.valueOf(changedTask.getCreatedAt()), Timestamp.valueOf(changedTask.getUpdatedAt()),
-                        changedTask.getText(), changedTask.getStatus().toString(),
+                        changedTask.getOwner(), changedTask.getText(), changedTask.getStatus().toString(),
                         Timestamp.valueOf(changedTask.getUpdatedAt())
                 );
                 if (rowsInsert > 0) {
@@ -227,15 +241,15 @@ public class DatabaseTaskRepository implements PaginationTaskRepository {
         return null;
     }
 
-    private Integer getResultSize(final String status) {
+    private Integer getResultSize(final String status, final String owner) {
         try {
             return jdbcOperations.queryForObject("SELECT count(*)" +
-                            "FROM task WHERE status = ?",
+                            "FROM task WHERE status = ? AND owner = ?",
                     (resultSet, i) -> {
                         int totalSize = Integer.valueOf(resultSet.getString(1));
                         return totalSize;
                     },
-                    status);
+                    status, owner);
         } catch (DataAccessException e) {
             logger.error(e.getMessage());
             return -1;
